extracted dsl methods into modules
