+++ lib/grape/validations.rb
@@ -84,6 +84,8 @@ module Grape
     class ParamsScope
       attr_accessor :element, :parent
 
+      include Grape::DSL::Parameters
+
       def initialize(opts, &block)
         @element  = opts[:element]
         @parent   = opts[:parent]
@@ -102,76 +104,6 @@ module Grape
         return true if parent.nil?
         parent.should_validate?(parameters)
       end
-
-      def requires(*attrs, &block)
-        orig_attrs = attrs.clone
-
-        opts = attrs.last.is_a?(Hash) ? attrs.pop : nil
-
-        if opts && opts[:using]
-          require_required_and_optional_fields(attrs.first, opts)
-        else
-          validate_attributes(attrs, opts, &block)
-
-          block_given? ? new_scope(orig_attrs, &block) :
-            push_declared_params(attrs)
-        end
-      end
-
-      def optional(*attrs, &block)
-        orig_attrs = attrs
-
-        validations = {}
-        validations.merge!(attrs.pop) if attrs.last.is_a?(Hash)
-        validations[:type] ||= Array if block_given?
-        validates(attrs, validations)
-
-        block_given? ? new_scope(orig_attrs, true, &block) :
-          push_declared_params(attrs)
-      end
-
-      def mutually_exclusive(*attrs)
-        validates(attrs, mutual_exclusion: true)
-      end
-
-      def exactly_one_of(*attrs)
-        validates(attrs, exactly_one_of: true)
-      end
-
-      def at_least_one_of(*attrs)
-        validates(attrs, at_least_one_of: true)
-      end
-
-      def group(*attrs, &block)
-        requires(*attrs, &block)
-      end
-
-      def params(params)
-        params = @parent.params(params) if @parent
-        if @element
-          if params.is_a?(Array)
-            params = params.flat_map { |el| el[@element] || {} }
-          elsif params.is_a?(Hash)
-            params = params[@element] || {}
-          else
-            params = {}
-          end
-        end
-        params
-      end
-
-      def use(*names)
-        named_params = @api.settings[:named_params] || {}
-        options = names.last.is_a?(Hash) ? names.pop : {}
-        names.each do |name|
-          params_block = named_params.fetch(name) do
-            raise "Params :#{name} not found!"
-          end
-          instance_exec(options, &params_block)
-        end
-      end
-      alias_method :use_scope, :use
-      alias_method :includes, :use
 
       def full_name(name)
         return "#{@parent.full_name(@element)}[#{name}]" if @parent
@@ -293,27 +225,6 @@ module Grape
           (@api.settings.peek[:validations] ||= []) << validator_class.new(attrs, options, doc_attrs[:required], self)
         else
           raise Grape::Exceptions::UnknownValidator.new(type)
-        end
-      end
-    end
-
-    # This module is mixed into the API Class.
-    module ClassMethods
-      def reset_validations!
-        settings.peek[:declared_params] = []
-        settings.peek[:validations] = []
-      end
-
-      def params(&block)
-        ParamsScope.new(api: self, type: Hash, &block)
-      end
-
-      def document_attribute(names, opts)
-        @last_description ||= {}
-        @last_description[:params] ||= {}
-        Array(names).each do |name|
-          @last_description[:params][name[:full_name].to_s] ||= {}
-          @last_description[:params][name[:full_name].to_s].merge!(opts)
         end
       end
     end
